+++ src/components/skeleton/index.ts
@@ -1,1 +1,3 @@
 export * from "./skeleton";
+export * from "./skeleton-circle";
+export * from "./skeleton-text";