+++ src/components/skeleton/skeleton.tsx
@@ -1,13 +1,18 @@
-import { mergeProps, onMount, Show, splitProps } from "solid-js";
+import { Property } from "csstype";
+import { createEffect, createSignal, mergeProps, on, Show, splitProps } from "solid-js";
 
+import { fadeIn } from "@/styled-system/keyframes";
 import { ColorProps } from "@/styled-system/props/color";
+import { colorTokenToCssVar } from "@/theme/utils";
 import { classNames, createClassSelector } from "@/utils/css";
 
 import { Box } from "../box/box";
 import { ElementType, HTMLHopeProps } from "../types";
 import { skeletonStyles } from "./skeleton.styles";
-import { Property } from "csstype";
-import { fadeIn } from "@/styled-system/keyframes";
+
+/* -------------------------------------------------------------------------------------------------
+ * Skeleton
+ * -----------------------------------------------------------------------------------------------*/
 
 interface SkeletonOptions {
   /**
@@ -21,26 +26,29 @@ interface SkeletonOptions {
   endColor?: ColorProps["backgroundColor"];
 
   /**
-   * If `true`, it'll render its children with a nice fade transition.
-   */
-  loaded?: boolean;
-
-  /**
    * The animation speed in CSS unit.
    */
   speed?: Property.AnimationDuration;
 
   /**
-   * The fadeIn duration in CSS unit.
+   * The loaded children fadeIn animation duration in CSS unit.
    */
   fadeDuration?: Property.AnimationDuration;
+
+  /**
+   * If `true`, it'll render its children with a nice fade animation.
+   */
+  loaded?: boolean;
 }
 
 export type SkeletonProps<C extends ElementType = "div"> = HTMLHopeProps<C, SkeletonOptions>;
 
-const hopeSkeletonClass = "hope-skeleton";
+export const hopeSkeletonClass = "hope-skeleton";
 
 export function Skeleton<C extends ElementType = "div">(props: SkeletonProps<C>) {
+  // Animation when the content is loaded
+  const [loadedAnimation, setLoadedAnimation] = createSignal("none");
+
   const defaultProps: SkeletonProps<"div"> = {
     speed: "800ms",
     fadeDuration: "400ms",
@@ -49,7 +57,6 @@ export function Skeleton<C extends ElementType = "div">(props: SkeletonProps<C>)
   const propsWithDefault: SkeletonProps<"div"> = mergeProps(defaultProps, props);
   const [local, others] = splitProps(propsWithDefault, [
     "class",
-    "children",
     "startColor",
     "endColor",
     "loaded",
@@ -61,43 +68,32 @@ export function Skeleton<C extends ElementType = "div">(props: SkeletonProps<C>)
     return classNames(
       local.class,
       hopeSkeletonClass,
-      skeletonStyles({
-        css: {
-          "--hope--startColor": colorTokenToCssVar(local.startColor ?? "$neutral2"),
-          "--hope--endColor": colorTokenToCssVar(local.endColor ?? "$neutral8"),
-          animationDuration: local.speed,
-        },
-      })
+      local.loaded
+        ? undefined
+        : skeletonStyles({
+            css: {
+              $$startColor: colorTokenToCssVar(local.startColor ?? "$neutral2"),
+              $$endColor: colorTokenToCssVar(local.endColor ?? "$neutral8"),
+              animationDuration: local.speed,
+            },
+          })
     );
   };
 
-  const loadedSkeletonClasses = () => {
-    return classNames(local.class, hopeSkeletonClass);
-  };
-
-  const loadedSkeletonAnimation = () => {
-    return `${fadeIn()} ${local.fadeDuration}`;
-  };
+  // The animation only applies when `loaded` goes from false to true.
+  createEffect(
+    on(
+      () => local.loaded,
+      (_, prev) => setLoadedAnimation(prev === true ? "none" : `${fadeIn()} ${local.fadeDuration}`),
+      { defer: true }
+    )
+  );
 
   return (
     <Show when={local.loaded} fallback={<Box class={skeletonClasses()} {...others} />}>
-      <Box class={loadedSkeletonClasses()} animation={loadedSkeletonAnimation()}>
-        {local.children}
-      </Box>
+      <Box class={skeletonClasses()} animation={loadedAnimation()} {...others} />
     </Show>
   );
 }
 
 Skeleton.toString = () => createClassSelector(hopeSkeletonClass);
-
-/**
- * Return the css variable associated with the given token if exists, or the token itself otherwise.
- */
-function colorTokenToCssVar(token: string): string {
-  if (!token.startsWith("$")) {
-    return token;
-  }
-
-  // ex: "$primary9" -> "var(--hope-colors-primary9)"
-  return `var(--hope-colors-${token.substring(1)})`;
-}