+++ src/components/stack/stack.styles.ts
@@ -1,0 +1,5 @@
+import { css } from "@/styled-system/stitches.config";
+
+export const stackStyles = css({
+  display: "flex",
+});